+++ trunk/dsol-core/src/main/java/nl/tudelft/simulation/jstats/distributions/DistGeometric.java
@@ -57,7 +57,7 @@ public class DistGeometric extends DistDiscrete
 
     /** {@inheritDoc} */
     @Override
-    public double probability(final int observation)
+    public double probability(final long observation)
     {
         if (observation >= 0)
         {
